move get_importance_traces to utils
